Move Windows backend test hooks to tests
Remove in-tree test hooks and the LIBVIRTUALHID_ENABLE_TEST_HOOKS build define, and relocate Windows backend test-only code into the test fixtures. Changes: removed the test-hook compile definition from src/CMakeLists.txt and removed the guarded test hooks from src/platform/windows/windows_backend.cpp; renamed header into tests/fixtures/include/fixtures/windows_backend_test_hooks.hpp; added tests/fixtures/windows_backend_test_hooks.cpp (defines test hook implementations and includes the backend implementation under a test-only macro); updated tests/CMakeLists.txt to compile the new fixture source and adjusted test include in tests/unit/test_windows_backend.cpp. This keeps test-only hooks out of the production library build and centralizes test fixtures in the tests tree.
